Hotfix: Adding an extra title page as the first page of thesis (see long comment below)
Based on the Central Library style, thesis paper-based pages starts with
an optional page resembling "In the name of God" following the thesis page.
After bookbinding this results to this stylish order:
	- Golden cover showing the thesis title
	- In the name of God on first paper-based page
	- The thesis title (again) on second paper-based page

Of course this style was not supposed to be followed when students print temporary
versions (without bookbinding and without any cover page). To fix this bug
and let students in hurry to deliver the temporary version of thesis without
thinking about putting the thesis title on top of other pages! this hotfix
adds an additional title page as the first page.

However, you can (and it's better) to remove that extra page before the
final bookbinding (*it has no influence on page numbering*).